Moving defaults to version related folder preparing for data field changes
