+++ packages/optimise-core/src/db/available_fields_ce.table.js
@@ -1,4 +1,4 @@
-import ceFields from './defaults/ceFields.json';
+import ceFields from './defaults_v1/ceFields.json';
 
 export const TABLE_NAME = 'AVAILABLE_FIELDS_CE';
 export const PRIORITY = 1;
