+++ packages/optimise-core/src/db/defaults_v1/drugs.json
