+++ packages/optimise-core/src/db/ethnicity.table.js
@@ -1,4 +1,4 @@
-import ethnicities from './defaults/ethnicities.json';
+import ethnicities from './defaults_v1/ethnicities.json';
 
 export const TABLE_NAME = 'ETHNICITIES';
 export const PRIORITY = 0;
