+++ packages/optimise-core/src/db/pregnancy_outcomes.table.js
@@ -1,4 +1,4 @@
-import pregnancyOutcomes from './defaults/pregnancyOutcomes.json';
+import pregnancyOutcomes from './defaults_v1/pregnancyOutcomes.json';
 
 export const TABLE_NAME = 'PREGNANCY_OUTCOMES';
 export const PRIORITY = 0;
